requirements.txt
flask==3.1.0
python-dotenv==1.0.1
pandas==2.2.3
numpy==2.0.2
google-generativeai==0.8.3
werkzeug==3.1.3
tabulate==0.9.0
pytest==8.3.4
